refactor: group files by domain rather than technology
