+++ src/SidePanel/OtherSettingsView.jsx
