+++ src/SidePanel/PacketView.jsx
