+++ src/SidePanel/ToggleChannelModeView.jsx
