+++ src/SidePanel/ToggleTestModeView.jsx
