+++ src/SidePanel/transmitSetupView.js
@@ -36,15 +36,19 @@
 
 import { connect } from 'react-redux';
 
-import ChartView from '../components/ChartView';
+import * as SettingsActions from '../actions/settingsActions';
+import TransmitSetupView from './TransmitSetupView';
 
-export default connect((state, props) => ({
-    ...props,
-    selectedTestMode: state.app.settings.testMode,
-    currentChannel: state.app.test.currentChannel,
-    lastChannel: state.app.test.lastChannel,
-    lastReceived: state.app.test.lastReceived,
-    isRunning: state.app.test.isRunning,
-    txPower: state.app.settings.txPower,
-    update: state.app.test.update,
-}))(ChartView);
+export default connect(
+    (state, props) => ({
+        ...props,
+        txPowerIdx: state.app.settings.txPower,
+        boardType: state.app.device.board,
+        isRunning: state.app.test.isRunning,
+    }),
+    (dispatch, props) => ({
+        ...props,
+        txPowerUpdated: value =>
+            dispatch(SettingsActions.txPowerUpdated(value)),
+    })
+)(TransmitSetupView);
